requirements.txt
# Requirements untuk Nusantara Food Watch - PIHPS Scraper

# Core libraries
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Database (pilih salah satu atau semua)
psycopg2-binary==2.9.9  # PostgreSQL
sqlalchemy==2.0.23       # ORM

# Data visualization (untuk dashboard nanti)
streamlit==1.29.0
plotly==5.18.0
folium==0.15.1

# Utilities
python-dotenv==1.0.0    # Environment variables
schedule==1.2.0         # Job scheduling
pytz==2023.3           # Timezone handling

# Testing & Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0         # Code formatter
flake8==6.1.0          # Linter